refactor: dropdown component & orderBy by Hygraph
