+++ src/components/HomeMenu.tsx
@@ -1,10 +1,9 @@
 'use client';
 
 import { GlobalContext } from '@/contexts/GlobalContext';
-import * as DropdownMenu from '@radix-ui/react-dropdown-menu';
 import { useContext } from 'react';
-import { FaChevronDown } from 'react-icons/fa'
 import { styled } from 'styled-components';
+import  { OrderByDropdown } from './OrderByDropdown';
 
 const NavContainer = styled.nav`
     width: 100%;
@@ -56,70 +55,10 @@ const NavItems = styled.div`
     }
 `;
 
-const HoverMenu = styled(DropdownMenu.Trigger)`
-    display: flex;
-    align-items: center;
-    background: none;
-
-    border: none;
-    outline: none;
-    cursor: pointer;
-
-    color: var(--color-gray-600);
-    text-align: center;
-    font-size: 14px;
-
-    transition: transform 0.3s;
-
-    span {
-        margin-right: 8px;
-    }
-
-    &:hover {
-        color: var(--color-gray-800);
-    }
-
-    &[data-state="open"] {
-        svg {
-            transform: rotate(180deg);
-            transition: transform 0.3s;
-        }
-    }
-`;
-
-const DropdownContent = styled(DropdownMenu.Content)`
-    background-color: white;
-    padding: 12px 16px;
-    border-radius: 4px;
-
-    color: var(--color-gray-600);
-    line-height: 22px;
-    font-size: 14px;
-    
-    div {
-        margin-bottom: 4px;
-        cursor: pointer;
-
-        &:last-child {
-            margin-bottom: 0;
-        }
-
-        &:hover {
-            outline: none;
-            color: var(--color-gray-800);
-        }
-    }
-`
-
 export function HomeMenu() {
     const { 
         categoryValue,
         setCategoryValue,
-        orderBySelection,
-        setOrderBySelection,
-
-        orderBy,
-        setOrderBy
     } = useContext(GlobalContext);
 
     return (
@@ -147,35 +86,7 @@ export function HomeMenu() {
                     </span>
                 </NavItems>
 
-
-                <DropdownMenu.Root>
-                    <HoverMenu>
-                        <span>Organizar por</span>
-                        <FaChevronDown size={10} />
-                    </HoverMenu>
-
-                    <DropdownMenu.Portal>
-                        <DropdownContent>
-                            <DropdownMenu.Item onClick={() => setOrderBy({ orderItem: "created_at", direction: "DESC" })}>
-                                Novidades
-                            </DropdownMenu.Item>
-
-                            <DropdownMenu.Item onClick={() => setOrderBy({ orderItem: "price_in_cents", direction: "DESC" })}>
-                                Preço: Maior - menor
-                            </DropdownMenu.Item>
-
-                            <DropdownMenu.Item onClick={() => setOrderBy({ orderItem: "price_in_cents", direction: "ASC" })}>
-                                Preço: Menor - maior
-                            </DropdownMenu.Item>
-
-                            <DropdownMenu.Item onClick={() => setOrderBy({ orderItem: "sales", direction: "DESC" })}>
-                                Mais vendidos
-                            </DropdownMenu.Item>
-
-                        </DropdownContent>
-                    </DropdownMenu.Portal>
-                </DropdownMenu.Root>
-
+                <OrderByDropdown />
             </NavRow>
         </NavContainer>
     )
